Save command test cases
